Instantiate RateLimitAdminSite in apps.py
We have seen increasing cases of the bug described in
https://stackoverflow.com/questions/62536538/django-runserver-failing-during-admin-site-checks
when running the https://github.com/edx/edx-platform development
environment. This applies the recommended fix to see if that resolves
our issues.
