+++ ratelimitbackend/__init__.py
@@ -1,1 +1,3 @@
-__version__ = '2.0'
+__version__ = '2.0.1'
+
+default_app_config = 'ratelimitbackend.apps.DjangoRatelimitBackendConfig'
